refactor: add media properties
